requirements.txt
flask
nanoid
mysql-connector-python
redis
PyJWT
cryptography
bcrypt
validators
celery
prometheus-client
geoip2
pyyaml
ua-parser 
user-agents
python-dotenv


pytest
pytest-mock
requests